added conditional check before creating symlink

+++ setup.sh
@@ -14,7 +14,10 @@ mkdir -p $HOME/bin
 for i in $ORG_DIR/bin/*
 do
    f=`basename $i`
-   ln -s $ORG_DIR/bin/$f $HOME/bin/$f
+   if [ ! -f $HOME/bin/$f ]
+   then
+      ln -s $ORG_DIR/bin/$f $HOME/bin/$f
+   fi
 done
 
 mkdir -p ~/tmp/backup
